+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ##### Enhancements
 
+* Add support for `Configurable` types when dequeuing reusable views.
+[Will McGinty](https://github.com/willmcginty)
+[#53](https://github.com/BottleRocketStudios/iOS-UtiliKit/pull/58)
+
 ##### Bug Fixes
 
-* None.
+* None
 
 
 ## 1.4.0 (2019-04-30)